send mail
send mail to user by Brevo
