+++ api-gateway/src/main/java/com/devteria/gateway/configuration/AuthenticationFilter.java
@@ -37,7 +37,8 @@ public class AuthenticationFilter implements GlobalFilter, Ordered {
 
     @NonFinal
     private String[] publicEndpoints = {"/identity/auth/.*",
-                                        "/identity/users/registration"};
+                                        "/identity/users/registration",
+                                        "/notification/email/send"};
 
     @Value("${app.api-prefix}")
     @NonFinal
