+++ notification-service/src/main/java/com/devteria/notification/controller/EmailController.java
@@ -1,0 +1,28 @@
+package com.devteria.notification.controller;
+
+import com.devteria.notification.dto.ApiResponse;
+import com.devteria.notification.dto.request.SendEmailRequest;
+import com.devteria.notification.dto.response.EmailResponse;
+import com.devteria.notification.service.EmailService;
+import lombok.AccessLevel;
+import lombok.RequiredArgsConstructor;
+import lombok.experimental.FieldDefaults;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+/**
+ * @author DatNv
+ */
+@RestController
+@RequiredArgsConstructor
+@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
+public class EmailController {
+    EmailService emailService;
+
+    @PostMapping("/email/send")
+    public ApiResponse<EmailResponse> sendEmail(@RequestBody SendEmailRequest request) {
+
+        return ApiResponse.<EmailResponse>builder().result(emailService.sendEmail(request)).build();
+    }
+}
