+++ notification-service/src/main/java/com/devteria/notification/dto/request/EmailRequest.java
@@ -1,0 +1,18 @@
+package com.devteria.notification.dto.request;
+
+import lombok.*;
+import lombok.experimental.FieldDefaults;
+
+import java.util.List;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class EmailRequest {
+    Sender sender;
+    List<Recipient> to;
+    String subject;
+    String htmlContent;
+}
