+++ notification-service/src/main/java/com/devteria/notification/dto/request/Recipient.java
@@ -1,0 +1,14 @@
+package com.devteria.notification.dto.request;
+
+import lombok.*;
+import lombok.experimental.FieldDefaults;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class Recipient {
+    String name;
+    String email;
+}
