+++ notification-service/src/main/java/com/devteria/notification/dto/request/SendEmailRequest.java
@@ -1,0 +1,15 @@
+package com.devteria.notification.dto.request;
+
+import lombok.*;
+import lombok.experimental.FieldDefaults;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class SendEmailRequest {
+    Recipient to;
+    String subject;
+    String htmlContent;
+}
