+++ notification-service/src/main/java/com/devteria/notification/dto/response/EmailResponse.java
@@ -1,0 +1,16 @@
+package com.devteria.notification.dto.response;
+
+import lombok.*;
+import lombok.experimental.FieldDefaults;
+
+/**
+ * @author DatNv
+ */
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class EmailResponse {
+    String messageId;
+}
