+++ notification-service/src/main/java/com/devteria/notification/exception/AppException.java
@@ -1,0 +1,19 @@
+package com.devteria.notification.exception;
+
+public class AppException extends RuntimeException {
+
+    public AppException(ErrorCode errorCode) {
+        super(errorCode.getMessage());
+        this.errorCode = errorCode;
+    }
+
+    private ErrorCode errorCode;
+
+    public ErrorCode getErrorCode() {
+        return errorCode;
+    }
+
+    public void setErrorCode(ErrorCode errorCode) {
+        this.errorCode = errorCode;
+    }
+}
