+++ notification-service/src/main/java/com/devteria/notification/repository/httpClient/EmailClient.java
@@ -1,0 +1,15 @@
+package com.devteria.notification.repository.httpClient;
+
+import com.devteria.notification.dto.request.EmailRequest;
+import com.devteria.notification.dto.response.EmailResponse;
+import org.springframework.cloud.openfeign.FeignClient;
+import org.springframework.http.MediaType;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
+
+@FeignClient(name = "email-client", url = "https://api.brevo.com")
+public interface EmailClient {
+    @PostMapping(value = "/v3/smtp/email", produces = MediaType.APPLICATION_JSON_VALUE)
+    EmailResponse sendEmail(@RequestHeader("api-key") String apiKey, @RequestBody EmailRequest request);
+}
