+++ notification-service/src/main/java/com/devteria/notification/service/EmailService.java
@@ -1,0 +1,42 @@
+package com.devteria.notification.service;
+
+import com.devteria.notification.dto.request.EmailRequest;
+import com.devteria.notification.dto.request.SendEmailRequest;
+import com.devteria.notification.dto.request.Sender;
+import com.devteria.notification.dto.response.EmailResponse;
+import com.devteria.notification.exception.AppException;
+import com.devteria.notification.exception.ErrorCode;
+import com.devteria.notification.repository.httpClient.EmailClient;
+import feign.FeignException;
+import lombok.AccessLevel;
+import lombok.RequiredArgsConstructor;
+import lombok.experimental.FieldDefaults;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+/**
+ * @author DatNv
+ */
+
+@Service
+@RequiredArgsConstructor
+@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
+public class EmailService {
+    EmailClient emailClient;
+    String apiKey = "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";
+
+    public EmailResponse sendEmail(SendEmailRequest request) {
+        EmailRequest emailRequest = EmailRequest.builder()
+                .sender(Sender.builder().name("Microservice").email("datnv11.fpoly@gmail.com").build())
+                .to(List.of(request.getTo()))
+                .subject(request.getSubject())
+                .htmlContent(request.getHtmlContent())
+                .build();
+        try {
+            return emailClient.sendEmail(apiKey, emailRequest);
+        } catch (FeignException e) {
+            throw new AppException(ErrorCode.CAN_NOT_SEND_EMAIL);
+        }
+    }
+}
