--- notification-service/src/main/java/com/devteria/profile/NotificationServiceApplication.java
@@ -1,12 +1,0 @@
-package com.devteria.profile;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class NotificationServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(NotificationServiceApplication.class, args);
-    }
-}
