+++ profile-service/src/main/java/com/devteria/notification/ProfileServiceApplication.java
@@ -1,0 +1,12 @@
+package com.devteria.notification;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ProfileServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ProfileServiceApplication.class, args);
+    }
+}
