+++ profile-service/src/main/java/com/devteria/notification/controller/InternalUserProfileController.java
@@ -1,0 +1,23 @@
+package com.devteria.notification.controller;
+
+import com.devteria.notification.dto.request.ProfileCreationRequest;
+import com.devteria.notification.dto.response.UserProfileResponse;
+import com.devteria.notification.service.UserProfileService;
+import lombok.AccessLevel;
+import lombok.RequiredArgsConstructor;
+import lombok.experimental.FieldDefaults;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
+public class InternalUserProfileController {
+    UserProfileService userProfileService;
+
+    @PostMapping("/internal/users")
+    UserProfileResponse createProfile(@RequestBody ProfileCreationRequest request) {
+        return userProfileService.createProfile(request);
+    }
+}
