+++ profile-service/src/main/java/com/devteria/notification/dto/request/ProfileCreationRequest.java
@@ -1,0 +1,19 @@
+package com.devteria.notification.dto.request;
+
+import java.time.LocalDate;
+
+import lombok.*;
+import lombok.experimental.FieldDefaults;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class ProfileCreationRequest {
+    String userId;
+    String firstName;
+    String lastName;
+    LocalDate dob;
+    String city;
+}
