+++ profile-service/src/main/java/com/devteria/notification/entity/UserProfile.java
@@ -1,0 +1,33 @@
+package com.devteria.notification.entity;
+
+import java.time.LocalDate;
+
+import org.springframework.data.neo4j.core.schema.GeneratedValue;
+import org.springframework.data.neo4j.core.schema.Id;
+import org.springframework.data.neo4j.core.schema.Node;
+import org.springframework.data.neo4j.core.schema.Property;
+import org.springframework.data.neo4j.core.support.UUIDStringGenerator;
+
+import lombok.*;
+import lombok.experimental.FieldDefaults;
+
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@FieldDefaults(level = AccessLevel.PRIVATE)
+@Node("user_profile")
+public class UserProfile {
+    @Id
+    @GeneratedValue(generatorClass = UUIDStringGenerator.class)
+    String id;
+
+    @Property("userId")
+    String userId;
+
+    String firstName;
+    String lastName;
+    LocalDate dob;
+    String city;
+}
