+++ profile-service/src/main/java/com/devteria/notification/mapper/UserProfileMapper.java
@@ -1,0 +1,14 @@
+package com.devteria.notification.mapper;
+
+import org.mapstruct.Mapper;
+
+import com.devteria.notification.dto.request.ProfileCreationRequest;
+import com.devteria.notification.dto.response.UserProfileResponse;
+import com.devteria.notification.entity.UserProfile;
+
+@Mapper(componentModel = "spring")
+public interface UserProfileMapper {
+    UserProfile toUserProfile(ProfileCreationRequest request);
+
+    UserProfileResponse toUserProfileReponse(UserProfile entity);
+}
