+++ profile-service/src/main/java/com/devteria/notification/repository/UserProfileRepository.java
@@ -1,0 +1,9 @@
+package com.devteria.notification.repository;
+
+import org.springframework.data.neo4j.repository.Neo4jRepository;
+import org.springframework.stereotype.Repository;
+
+import com.devteria.notification.entity.UserProfile;
+
+@Repository
+public interface UserProfileRepository extends Neo4jRepository<UserProfile, String> {}
