+++ profile-service/src/main/java/com/devteria/notification/service/UserProfileService.java
@@ -1,0 +1,45 @@
+package com.devteria.notification.service;
+
+import org.springframework.stereotype.Service;
+
+import com.devteria.notification.dto.request.ProfileCreationRequest;
+import com.devteria.notification.dto.response.UserProfileResponse;
+import com.devteria.notification.entity.UserProfile;
+import com.devteria.notification.mapper.UserProfileMapper;
+import com.devteria.notification.repository.UserProfileRepository;
+
+import lombok.AccessLevel;
+import lombok.RequiredArgsConstructor;
+import lombok.experimental.FieldDefaults;
+import lombok.extern.slf4j.Slf4j;
+
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
+@Slf4j
+public class UserProfileService {
+    UserProfileRepository userProfileRepository;
+    UserProfileMapper userProfileMapper;
+
+    public UserProfileResponse createProfile(ProfileCreationRequest request) {
+        UserProfile userProfile = userProfileMapper.toUserProfile(request);
+        userProfile = userProfileRepository.save(userProfile);
+
+        return userProfileMapper.toUserProfileReponse(userProfile);
+    }
+
+    public UserProfileResponse getProfile(String id) {
+        UserProfile userProfile =
+                userProfileRepository.findById(id).orElseThrow(() -> new RuntimeException("Profile not found"));
+
+        return userProfileMapper.toUserProfileReponse(userProfile);
+    }
+
+    public List<UserProfileResponse> getAllProfiles() {
+        var profiles = userProfileRepository.findAll();
+
+        return profiles.stream().map(userProfileMapper::toUserProfileReponse).toList();
+    }
+}
