--- profile-service/src/main/java/com/devteria/profile/ProfileServiceApplication.java
@@ -1,12 +1,0 @@
-package com.devteria.profile;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ProfileServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ProfileServiceApplication.class, args);
-    }
-}
