--- profile-service/src/main/java/com/devteria/profile/dto/ApiResponse.java
@@ -1,19 +1,0 @@
-package com.devteria.profile.dto;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.*;
-import lombok.experimental.FieldDefaults;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@FieldDefaults(level = AccessLevel.PRIVATE)
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class ApiResponse<T> {
-    @Builder.Default
-    private int code = 1000;
-
-    private String message;
-    private T result;
-}
