--- profile-service/src/main/java/com/devteria/profile/dto/response/UserProfileResponse.java
@@ -1,19 +1,0 @@
-package com.devteria.profile.dto.response;
-
-import java.time.LocalDate;
-
-import lombok.*;
-import lombok.experimental.FieldDefaults;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class UserProfileResponse {
-    String id;
-    String firstName;
-    String lastName;
-    LocalDate dob;
-    String city;
-}
