--- profile-service/src/main/java/com/devteria/profile/exception/AppException.java
@@ -1,19 +1,0 @@
-package com.devteria.profile.exception;
-
-public class AppException extends RuntimeException {
-
-    public AppException(ErrorCode errorCode) {
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-    }
-
-    private ErrorCode errorCode;
-
-    public ErrorCode getErrorCode() {
-        return errorCode;
-    }
-
-    public void setErrorCode(ErrorCode errorCode) {
-        this.errorCode = errorCode;
-    }
-}
