--- profile-service/src/main/java/com/devteria/profile/mapper/UserProfileMapper.java
@@ -1,14 +1,0 @@
-package com.devteria.profile.mapper;
-
-import org.mapstruct.Mapper;
-
-import com.devteria.profile.dto.request.ProfileCreationRequest;
-import com.devteria.profile.dto.response.UserProfileResponse;
-import com.devteria.profile.entity.UserProfile;
-
-@Mapper(componentModel = "spring")
-public interface UserProfileMapper {
-    UserProfile toUserProfile(ProfileCreationRequest request);
-
-    UserProfileResponse toUserProfileReponse(UserProfile entity);
-}
