--- profile-service/src/main/java/com/devteria/profile/repository/UserProfileRepository.java
@@ -1,9 +1,0 @@
-package com.devteria.profile.repository;
-
-import org.springframework.data.neo4j.repository.Neo4jRepository;
-import org.springframework.stereotype.Repository;
-
-import com.devteria.profile.entity.UserProfile;
-
-@Repository
-public interface UserProfileRepository extends Neo4jRepository<UserProfile, String> {}
