--- profile-service/src/main/java/com/devteria/profile/service/UserProfileService.java
@@ -1,45 +1,0 @@
-package com.devteria.profile.service;
-
-import org.springframework.stereotype.Service;
-
-import com.devteria.profile.dto.request.ProfileCreationRequest;
-import com.devteria.profile.dto.response.UserProfileResponse;
-import com.devteria.profile.entity.UserProfile;
-import com.devteria.profile.mapper.UserProfileMapper;
-import com.devteria.profile.repository.UserProfileRepository;
-
-import lombok.AccessLevel;
-import lombok.RequiredArgsConstructor;
-import lombok.experimental.FieldDefaults;
-import lombok.extern.slf4j.Slf4j;
-
-import java.util.List;
-
-@Service
-@RequiredArgsConstructor
-@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
-@Slf4j
-public class UserProfileService {
-    UserProfileRepository userProfileRepository;
-    UserProfileMapper userProfileMapper;
-
-    public UserProfileResponse createProfile(ProfileCreationRequest request) {
-        UserProfile userProfile = userProfileMapper.toUserProfile(request);
-        userProfile = userProfileRepository.save(userProfile);
-
-        return userProfileMapper.toUserProfileReponse(userProfile);
-    }
-
-    public UserProfileResponse getProfile(String id) {
-        UserProfile userProfile =
-                userProfileRepository.findById(id).orElseThrow(() -> new RuntimeException("Profile not found"));
-
-        return userProfileMapper.toUserProfileReponse(userProfile);
-    }
-
-    public List<UserProfileResponse> getAllProfiles() {
-        var profiles = userProfileRepository.findAll();
-
-        return profiles.stream().map(userProfileMapper::toUserProfileReponse).toList();
-    }
-}
